se modifico la agenda para llamar los salones desde classroomservice
